requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.18.20
botocore==1.21.20
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Fabric3==1.14.post1
gunicorn==20.1.0
idna==3.3
invoke==1.6.0
jmespath==0.10.0
mysqlclient==2.0.3
oauthlib==3.1.1
paramiko==2.7.2
pathlib2==2.3.6
Pillow==8.3.1
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
